+++ src/defaults.ts
@@ -7,4 +7,5 @@ export const defaults: Options = {
 	headerButtonClass: '',
 	animate: true,
 	multiselect: true,
+	dataAttributePrefix: 'accordion',
 };