+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Added
 
 - Added `headerButtonClass` option.
+- Added `dataAttributePrefix` option.
 - Added `id` property to the Accordion instance.
 - Added `root` property to the Accordion instance.
 - Added `items` property to the Accordion instance.
@@ -17,6 +18,7 @@
 
 ### Changed
 
+- Data attributes are now prefixed with `accordion`.
 - Improved type information.
 - `initialisedClass`, `activePanelClass` and `activeHeaderClass` are no longer added by default.
 - Event names passed to the `on` and `off` methods are no longer prefixed with `accordion:`.